Move content of schook/simsync.lua to simsync.lua
